Make aggregate_cpu_speed a sql accessible column
Aggregate hardware knows the particulars of this column.
Instead, define it on the hardware class and access it as such.

aggregate_cpu_speed was a little tricky to get right when either
the aggregate column was brought back OR the individual columns were
brought back. Added a spec in aggregate_mixin to ensure someone did not
optimize this safeguard away.

The new sql for aggregates does multiply, so is a hint less optimized
but opens us up for scopes