update event_stream timeline indexes
most of the queries for the timeline are sorting by
timeline given another id. This is basicaly causing a
full tablescan

These indexes do increase the disk usage, but allow the indexes
to actually be used.

Queries may need to be tweaked since the copious OR clauses
encourage the tablescan.