Consolidate rbac parent and class filtering
We want to get away from hitting the database with 4 queries and
fetching a bunch of ids. Doing set logic in ruby.
Then sending the ids back to the database for a final query.
Especially since set logic is sql's forte.

Instead, a scope will be built up with the necessary filtering.
Most of the time, no extra querying will be performed

belongsto will still have a sub query 

This is the first step in a refactor, so param lists are longer