Remove taggings and polymorphic from references
We pass includes (and references) into rbac.
They both come in as includes with a hash of references.
In truth, references requires a flat array of table names.

In Rbac, we convert the references from this references hash to an array using `includes_to_references`.

This pr modifies `includes_to_references` to
- no longer include taggings
- no longer include virtual attributes